SSD Evaluation #615
enhancing display process of data from departments to evaluations
