+++ evaluation/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
             android:exported="false" />
         <activity
             android:name=".Activity.SSDD.Activity_SSDD_Evaluation"
-            android:exported="false" />
+            android:exported="false"/>
         <activity
             android:name=".Activity.PacitaReward.Activity_BranchRecord_Details"
             android:exported="false"
